[kissmanga] remove module
