+++ docs/gallery-dl.conf
@@ -88,10 +88,6 @@
             "highlights": false,
             "videos": true
         },
-        "kissmanga":
-        {
-            "captcha": "stop"
-        },
         "nijie":
         {
             "username": null,
@@ -114,10 +110,6 @@
         {
             "wait-min": 3.0,
             "wait-max": 6.0
-        },
-        "readcomiconline":
-        {
-            "captcha": "stop"
         },
         "reddit":
         {
