+++ gallery_dl/extractor/__init__.py
@@ -59,7 +59,6 @@ modules = [
     "kabeuchi",
     "keenspot",
     "khinsider",
-    "kissmanga",
     "komikcast",
     "konachan",
     "lineblog",
